refactor: split secret service to link, profile, aggregate dedicated service
