+++ src/main/java/org/phong/zenflow/plugin/services/PluginService.java
@@ -18,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
 
 @AllArgsConstructor
@@ -110,5 +111,9 @@ public class PluginService {
     public Plugin findPluginById(UUID id) {
         return pluginRepository.findById(id)
                 .orElseThrow(() -> new PluginException("Plugin not found with id: " + id));
+    }
+
+    public Optional<Plugin> findByKey(String pluginKey) {
+        return pluginRepository.findByKey(pluginKey);
     }
 }
