+++ src/main/java/org/phong/zenflow/secret/service/SecretService.java
@@ -7,37 +7,18 @@ import org.phong.zenflow.core.services.AuthService;
 import org.phong.zenflow.log.auditlog.annotations.AuditLog;
 import org.phong.zenflow.log.auditlog.enums.AuditAction;
 import org.phong.zenflow.project.service.ProjectService;
-import org.phong.zenflow.secret.dto.AggregatedSecretSetupDto;
-import org.phong.zenflow.secret.dto.CreateProfileSecretsRequest;
 import org.phong.zenflow.secret.dto.CreateSecretBatchRequest;
 import org.phong.zenflow.secret.dto.CreateSecretRequest;
-import org.phong.zenflow.secret.dto.LinkProfileToNodeRequest;
-import org.phong.zenflow.secret.dto.LinkSecretToNodeRequest;
-import org.phong.zenflow.secret.dto.NodeProfileLinkDto;
-import org.phong.zenflow.secret.dto.NodeSecretLinksDto;
-import org.phong.zenflow.secret.dto.ProfileSecretListDto;
 import org.phong.zenflow.secret.dto.SecretDto;
 import org.phong.zenflow.secret.dto.UpdateSecretRequest;
-import org.phong.zenflow.secret.enums.SecretScope;
 import org.phong.zenflow.secret.exception.SecretDomainException;
 import org.phong.zenflow.secret.exception.SecretNotFoundException;
 import org.phong.zenflow.secret.infrastructure.mapstruct.SecretMapper;
-import org.phong.zenflow.secret.infrastructure.persistence.entity.ProfileSecretLink;
 import org.phong.zenflow.secret.infrastructure.persistence.entity.Secret;
-import org.phong.zenflow.secret.infrastructure.persistence.entity.SecretNodeLink;
-import org.phong.zenflow.secret.infrastructure.persistence.entity.SecretProfile;
-import org.phong.zenflow.secret.infrastructure.persistence.entity.SecretProfileNodeLink;
-import org.phong.zenflow.secret.infrastructure.persistence.repository.SecretProfileRepository;
-import org.phong.zenflow.secret.infrastructure.persistence.repository.ProfileSecretLinkRepository;
-import org.phong.zenflow.secret.infrastructure.persistence.repository.SecretProfileNodeLinkRepository;
-import org.phong.zenflow.secret.infrastructure.persistence.repository.SecretNodeLinkRepository;
 import org.phong.zenflow.secret.infrastructure.persistence.repository.SecretRepository;
 import org.phong.zenflow.secret.util.AESUtil;
 import org.phong.zenflow.user.service.UserService;
 import org.phong.zenflow.workflow.infrastructure.persistence.repository.WorkflowRepository;
-import org.phong.zenflow.plugin.subdomain.node.service.PluginNodeService;
-import org.phong.zenflow.plugin.infrastructure.persistence.repository.PluginRepository;
-import org.phong.zenflow.plugin.services.PluginService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -55,20 +36,12 @@ import java.util.stream.Collectors;
 @Transactional
 public class SecretService {
     private final SecretRepository secretRepository;
-    private final SecretProfileRepository secretProfileRepository;
-    private final ProfileSecretLinkRepository profileSecretLinkRepository;
-    private final SecretProfileNodeLinkRepository secretProfileNodeLinkRepository;
-    private final SecretNodeLinkRepository secretNodeLinkRepository;
     private final SecretMapper secretMapper;
     private final AESUtil aesUtil;
     private final UserService userService;
     private final WorkflowRepository workflowRepository;
     private final ProjectService projectService;
     private final AuthService authService;
-    private final SecretProfileSchemaValidator validator;
-    private final PluginNodeService pluginNodeService;
-    private final PluginService pluginService;
-    private final PluginRepository pluginRepository;
 
     @Transactional(readOnly = true)
     public List<SecretDto> getAllSecrets() {
@@ -243,90 +216,6 @@ public class SecretService {
                 .collect(Collectors.toList());
     }
 
-    /**
-     * Retrieves all secrets for a workflow grouped by their profile name.
-     *
-     * @param workflowId the workflow identifier
-     * @return map keyed by profile name, each containing a map of secret key to decrypted value
-     */
-    public ProfileSecretListDto getProfileSecretMapByWorkflowId(UUID workflowId) {
-        Map<String, Map<String, String>> collect = getProfilesKeyMapByWorkflowId(workflowId);
-        return new ProfileSecretListDto(collect);
-    }
-
-    public ProfileSecretListDto createProfileSecrets(UUID workflowId, CreateProfileSecretsRequest request) {
-        if (request.isDuplicated()) {
-            throw new SecretDomainException("Duplicate keys found in the request.");
-        }
-
-        boolean exists = secretProfileRepository.existsByNameAndWorkflowIdAndPluginId(
-                request.getName(), workflowId, request.getPluginId());
-        if (exists) {
-            throw new SecretDomainException("Profile with name '" + request.getName() + "' already exists for this workflow.");
-        }
-
-        SecretProfile profile = new SecretProfile();
-        profile.setWorkflow(workflowRepository.getReferenceById(workflowId));
-        profile.setName(request.getName());
-        profile.setScope(SecretScope.WORKFLOW);
-        profile.setUser(authService.getReferenceCurrentUser());
-        profile.setPlugin(pluginService.findPluginById(request.getPluginId()));
-        if (request.getPluginNodeId() != null) {
-            profile.setPluginNode(pluginNodeService.findById(request.getPluginNodeId()));
-        }
-        profile = secretProfileRepository.save(profile);
-
-        SecretProfile finalProfile = profile;
-        List<Secret> secrets = request.getSecrets().entrySet().stream()
-                .map(entry -> {
-                    Secret secret = new Secret();
-                    secret.setWorkflow(finalProfile.getWorkflow());
-                    secret.setKey(entry.getKey());
-                    secret.setScope(SecretScope.WORKFLOW);
-
-                    try {
-                        secret.setEncryptedValue(aesUtil.encrypt(entry.getValue()));
-                    } catch (Exception e) {
-                        throw new RuntimeException(e);
-                    }
-                    secret.setUser(authService.getReferenceCurrentUser());
-                    return secret;
-                })
-                .collect(Collectors.toList());
-
-        boolean isValid = validator.validate(request.getPluginId(), request.getPluginNodeId(), request.getSecrets());
-        if (!isValid) {
-            throw new SecretDomainException("Secrets do not conform to the required schema!");
-        }
-
-        List<Secret> savedSecrets = secretRepository.saveAll(secrets);
-
-        // Link secrets to the created profile
-        var links = savedSecrets.stream().map(sec -> {
-            var link = new ProfileSecretLink();
-            link.setProfile(finalProfile);
-            link.setSecret(sec);
-            return link;
-        }).collect(Collectors.toList());
-        profileSecretLinkRepository.saveAll(links);
-
-        Map<String, Map<String, String>> profileMap = Map.of(
-                finalProfile.getName(),
-                savedSecrets.stream().collect(Collectors.toMap(
-                        Secret::getKey,
-                        s -> {
-                            try {
-                                return aesUtil.decrypt(s.getEncryptedValue());
-                            } catch (Exception e) {
-                                throw new SecretDomainException("Can't decrypt value for workflowId: " + workflowId, e);
-                            }
-                        }
-                ))
-        );
-
-        return new ProfileSecretListDto(profileMap);
-    }
-
     private SecretDto mapToDto(Secret secret) {
         try {
             SecretDto dto = secretMapper.toDto(secret);
@@ -338,87 +227,6 @@ public class SecretService {
             log.error("Failed to decrypt secret value for id {}: {}", secret.getId(), e.getMessage(), e);
             throw new SecretDomainException("Failed to decrypt secret value", e);
         }
-    }
-
-    public void linkProfileToNode(UUID workflowId, LinkProfileToNodeRequest request) {
-        var profile = secretProfileRepository.getReferenceById(request.profileId());
-        var link = secretProfileNodeLinkRepository.findByWorkflowIdAndNodeKey(workflowId, request.nodeKey())
-                .orElseGet(SecretProfileNodeLink::new);
-
-        link.setWorkflow(workflowRepository.getReferenceById(workflowId));
-        link.setNodeKey(request.nodeKey());
-        link.setProfile(profile);
-        secretProfileNodeLinkRepository.save(link);
-    }
-
-    public void linkSecretToNode(UUID workflowId, LinkSecretToNodeRequest request) {
-        var secret = secretRepository.getReferenceById(request.secretId());
-        var existing = secretNodeLinkRepository.findByWorkflowIdAndNodeKeyAndSecretId(workflowId, request.nodeKey(), request.secretId());
-        if (existing.isPresent()) {
-            return; // already linked
-        }
-        var link = new SecretNodeLink();
-        link.setWorkflow(workflowRepository.getReferenceById(workflowId));
-        link.setNodeKey(request.nodeKey());
-        link.setSecret(secret);
-        secretNodeLinkRepository.save(link);
-    }
-
-    @Transactional(readOnly = true)
-    public boolean isProfileLinked(UUID workflowId, String nodeKey) {
-        return secretProfileNodeLinkRepository.findByWorkflowIdAndNodeKey(workflowId, nodeKey).isPresent();
-    }
-
-    @Transactional(readOnly = true)
-    public NodeProfileLinkDto getProfileLink(UUID workflowId, String nodeKey) {
-        return secretProfileNodeLinkRepository.findByWorkflowIdAndNodeKey(workflowId, nodeKey)
-                .map(link -> new NodeProfileLinkDto(
-                        workflowId,
-                        nodeKey,
-                        link.getProfile().getId()
-                ))
-                .orElse(null);
-    }
-
-    public void unlinkProfileFromNode(UUID workflowId, String nodeKey) {
-        secretProfileNodeLinkRepository.deleteByWorkflowIdAndNodeKey(workflowId, nodeKey);
-    }
-
-    @Transactional(readOnly = true)
-    public NodeSecretLinksDto getSecretLinks(UUID workflowId, String nodeKey) {
-        var ids = secretNodeLinkRepository.findByWorkflowIdAndNodeKey(workflowId, nodeKey)
-                .stream()
-                .map(link -> link.getSecret().getId())
-                .collect(Collectors.toList());
-        return new NodeSecretLinksDto(workflowId, nodeKey, ids);
-    }
-
-    public void unlinkSecretFromNode(UUID workflowId, String nodeKey, UUID secretId) {
-        secretNodeLinkRepository.deleteByWorkflowIdAndNodeKeyAndSecretId(workflowId, nodeKey, secretId);
-    }
-
-    public void unlinkAllSecretsFromNode(UUID workflowId, String nodeKey) {
-        secretNodeLinkRepository.deleteByWorkflowIdAndNodeKey(workflowId, nodeKey);
-    }
-
-    @Transactional(readOnly = true)
-    public List<UUID> getMissingSecretLinks(UUID workflowId, String nodeKey, List<UUID> expectedSecretIds) {
-        if (workflowId == null || nodeKey == null || expectedSecretIds == null || expectedSecretIds.isEmpty()) {
-            return List.of();
-        }
-        var linkedIds = secretNodeLinkRepository.findByWorkflowIdAndNodeKey(workflowId, nodeKey)
-                .stream()
-                .map(link -> link.getSecret().getId())
-                .collect(Collectors.toSet());
-
-        return expectedSecretIds.stream()
-                .filter(id -> id != null && !linkedIds.contains(id))
-                .toList();
-    }
-
-    @Transactional(readOnly = true)
-    public boolean areSecretsLinked(UUID workflowId, String nodeKey, List<UUID> requiredSecretIds) {
-        return getMissingSecretLinks(workflowId, nodeKey, requiredSecretIds).isEmpty();
     }
 
     @Transactional(readOnly = true)
@@ -438,14 +246,6 @@ public class SecretService {
     }
 
     @Transactional(readOnly = true)
-    public List<UUID> getLinkedSecretIds(UUID workflowId, String nodeKey) {
-        return secretNodeLinkRepository.findByWorkflowIdAndNodeKey(workflowId, nodeKey)
-                .stream()
-                .map(link -> link.getSecret().getId())
-                .collect(Collectors.toList());
-    }
-
-    @Transactional(readOnly = true)
     public Map<String, String> getSecretsKeyMapByWorkflowId(UUID workflowId) {
         return secretRepository.findByWorkflowId(workflowId)
                 .stream()
@@ -460,129 +260,5 @@ public class SecretService {
                         },
                         (existing, replacement) -> replacement
                 ));
-    }
-
-    @Transactional(readOnly = true)
-    public Map<String, Map<String, String>> getProfilesKeyMapByWorkflowId(UUID workflowId) {
-        return profileSecretLinkRepository.findByWorkflowId(workflowId)
-                .stream()
-                .collect(Collectors.groupingBy(
-                        link -> link.getProfile().getName(),
-                        Collectors.toMap(
-                                link -> link.getSecret().getKey(),
-                                link -> {
-                                    try {
-                                        return aesUtil.decrypt(link.getSecret().getEncryptedValue());
-                                    } catch (Exception e) {
-                                        throw new SecretDomainException("Can't decrypt value for workflowId: " + workflowId, e);
-                                    }
-                                },
-                                (existing, replacement) -> replacement
-                        )
-                ));
-    }
-
-    @Transactional(readOnly = true)
-    public AggregatedSecretSetupDto getAggregatedSecretsProfilesAndNodeIndex(UUID workflowId) {
-        // Secrets keyed by secretId to avoid collisions on duplicate keys
-        Map<String, String> secretsById = secretRepository.findByWorkflowId(workflowId)
-                .stream()
-                .collect(Collectors.toMap(
-                        s -> s.getId().toString(),
-                        s -> {
-                            try {
-                                return aesUtil.decrypt(s.getEncryptedValue());
-                            } catch (Exception e) {
-                                throw new SecretDomainException("Can't decrypt value for workflowId: " + workflowId, e);
-                            }
-                        },
-                        (existing, replacement) -> replacement
-                ));
-
-        // Secret id -> key for reconstructing per-node key maps
-        Map<String, String> secretKeys = secretRepository.findByWorkflowId(workflowId)
-                .stream()
-                .collect(Collectors.toMap(s -> s.getId().toString(), Secret::getKey, (a, b) -> b));
-
-        // Profiles keyed by profileId, values map secretKey -> decrypted value for that profile
-        Map<String, Map<String, String>> profilesById = profileSecretLinkRepository.findByWorkflowId(workflowId)
-                .stream()
-                .collect(Collectors.groupingBy(
-                        link -> link.getProfile().getId().toString(),
-                        Collectors.toMap(
-                                link -> link.getSecret().getKey(),
-                                link -> {
-                                    try {
-                                        return aesUtil.decrypt(link.getSecret().getEncryptedValue());
-                                    } catch (Exception e) {
-                                        throw new SecretDomainException("Can't decrypt value for workflowId: " + workflowId, e);
-                                    }
-                                },
-                                (existing, replacement) -> replacement
-                        )
-                ));
-
-        // Map nodeKey -> profileId (stable)
-        Map<String, String> nodeProfiles = secretProfileNodeLinkRepository.findAllByWorkflowId(workflowId)
-                .stream()
-                .collect(Collectors.toMap(
-                        SecretProfileNodeLink::getNodeKey,
-                        link -> link.getProfile().getId().toString(),
-                        (a, b) -> b
-                ));
-
-        // Map nodeKey -> [secretId] to reference stable identifiers
-        Map<String, List<String>> nodeSecrets = secretNodeLinkRepository.findByWorkflowId(workflowId)
-                .stream()
-                .collect(Collectors.groupingBy(
-                        SecretNodeLink::getNodeKey,
-                        Collectors.mapping(
-                                link -> link.getSecret().getId().toString(),
-                                Collectors.toList()
-                        )
-                ));
-
-        // Also expose profileId -> profileName for display purposes
-        Map<String, String> profileNames = secretProfileRepository.findAll().stream()
-                .filter(p -> p.getWorkflow() != null && workflowId.equals(p.getWorkflow().getId()))
-                .collect(Collectors.toMap(p -> p.getId().toString(), SecretProfile::getName, (a, b) -> b));
-
-        return new AggregatedSecretSetupDto(secretsById, profilesById, nodeProfiles, nodeSecrets, profileNames, secretKeys);
-    }
-
-    /**
-     * Checks whether a profile with the given name exists for a workflow and plugin key.
-     * This is used by validators to verify reference existence without exposing repository details.
-     */
-    @Transactional(readOnly = true)
-    public boolean profileExists(UUID workflowId, String pluginKey, String profileName) {
-        if (workflowId == null || pluginKey == null || profileName == null || profileName.isBlank()) {
-            return false;
-        }
-        return pluginRepository.findByKey(pluginKey)
-                .map(plugin -> secretProfileRepository.existsByNameAndWorkflowIdAndPluginId(profileName, workflowId, plugin.getId()))
-                .orElse(false);
-    }
-
-    @Transactional(readOnly = true)
-    public UUID resolveProfileId(UUID workflowId, String pluginKey, String profileName) {
-        if (workflowId == null || pluginKey == null || profileName == null || profileName.isBlank()) {
-            return null;
-        }
-        return pluginRepository.findByKey(pluginKey)
-                .flatMap(plugin -> secretProfileRepository.findByNameAndWorkflowIdAndPluginId(profileName, workflowId, plugin.getId()))
-                .map(SecretProfile::getId)
-                .orElse(null);
-    }
-
-    public void linkProfileIfExists(UUID workflowId, String nodeKey, String pluginKey, String profileName) {
-        try {
-            UUID profileId = resolveProfileId(workflowId, pluginKey, profileName);
-            if (profileId != null) {
-                linkProfileToNode(workflowId, new LinkProfileToNodeRequest(profileId, nodeKey));
-            }
-        } catch (Exception e) {
-            log.warn("Failed to link profile '{}' for node '{}' in workflow {}: {}", profileName, nodeKey, workflowId, e.getMessage());
-        }
     }
 }
