+++ src/main/java/org/phong/zenflow/secret/subdomain/link/dto/SecretNodeLinkInsertRequestDto.java
@@ -1,0 +1,8 @@
+package org.phong.zenflow.secret.subdomain.link.dto;
+
+import java.io.Serializable;
+import java.util.UUID;
+
+public record SecretNodeLinkInsertRequestDto(UUID secretId, String nodeKey) implements Serializable {
+
+}
