+++ src/main/java/org/phong/zenflow/secret/subdomain/link/listeners/SecretLinkSyncListener.java
@@ -1,7 +1,7 @@
-package org.phong.zenflow.secret.listeners;
+package org.phong.zenflow.secret.subdomain.link.listeners;
 
 import lombok.AllArgsConstructor;
-import org.phong.zenflow.secret.service.SecretLinkSyncService;
+import org.phong.zenflow.secret.subdomain.link.service.SecretLinkSyncService;
 import org.phong.zenflow.workflow.event.WorkflowDefinitionUpdatedEvent;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
@@ -10,7 +10,7 @@ import org.springframework.transaction.event.TransactionalEventListener;
 
 @Component
 @AllArgsConstructor
-public class SecretListener {
+public class SecretLinkSyncListener {
     private final SecretLinkSyncService linkSyncService;
 
     @Async
